Fixed some bugs and added a master template
The side bar works in single post page.
added a master template containing a head block and a content block.
The head block contains the links and styles which are repeated in all pages.
The name of user is shown in the contact and about us page if user is logged in.